+++ include/simplydt/datetime/date/comp/dt_vdate.hpp
